Community minor fixes

+++ src/main/java/data/cscl/Community.java
@@ -18,7 +18,6 @@ import java.util.Map.Entry;
 import java.util.Set;
 import java.util.TreeSet;
 
-import org.apache.log4j.BasicConfigurator;
 import org.apache.log4j.Logger;
 
 import data.AbstractDocument;
@@ -45,7 +44,6 @@ import services.discourse.CSCL.ParticipantEvaluation;
 import services.discourse.cohesion.CohesionGraph;
 import services.discourse.topicMining.TopicModeling;
 import services.replicatedWorker.SerialCorpusAssessment;
-import view.widgets.ReaderBenchView;
 import view.widgets.chat.ParticipantInvolvementView;
 import view.widgets.document.corpora.PaperConceptView;
 
@@ -267,8 +265,8 @@ public class Community extends AnalysisElement {
 				if (p.getSignificantInterventions().getBlocks().size() >= MIN_NO_CONTRIBUTIONS
 						&& noContentWords >= MIN_NO_CONTENT_WORDS) {
 					// build cohesion graph for additional indices
-
 					CohesionGraph.buildCohesionGraph(p.getSignificantInterventions());
+
 					for (IComplexityFactors f : complexityFactors) {
 						f.computeComplexityFactors(p.getSignificantInterventions());
 					}
@@ -514,18 +512,20 @@ public class Community extends AnalysisElement {
 		}
 	}
 
-	public static void processAllFolders(String folder, String prefix, String pathToLSA, String pathToLDA, Lang lang,
-			boolean usePOSTagging, boolean useTextualComplexity, Date startDate, Date endDate, int monthIncrement,
-			int dayIncrement) {
+	public static void processAllFolders(String folder, String prefix, boolean restartProcessing, String pathToLSA,
+			String pathToLDA, Lang lang, boolean usePOSTagging, boolean useTextualComplexity, Date startDate,
+			Date endDate, int monthIncrement, int dayIncrement) {
 		File dir = new File(folder);
 		if (dir.isDirectory()) {
 			File[] communityFolder = dir.listFiles();
 			for (File f : communityFolder) {
 				if (f.isDirectory() && f.getName().startsWith(prefix)) {
-					// remove checkpoint file
-					File checkpoint = new File(f + "/checkpoint.xml");
-					if (checkpoint.exists())
-						checkpoint.delete();
+					if (restartProcessing) {
+						// remove checkpoint file
+						File checkpoint = new File(f + "/checkpoint.xml");
+						if (checkpoint.exists())
+							checkpoint.delete();
+					}
 					SerialCorpusAssessment.processCorpus(f.getAbsolutePath(), pathToLSA, pathToLDA, lang, usePOSTagging,
 							true, null, null, true);
 					Community.processDocumentCollection(f.getAbsolutePath(), useTextualComplexity, startDate, endDate,
@@ -573,24 +573,5 @@ public class Community extends AnalysisElement {
 
 	public void setTimeframeSubCommunities(List<Community> timeframeSubCommunities) {
 		this.timeframeSubCommunities = timeframeSubCommunities;
-	}
-
-	public static void main(String[] args) {
-		BasicConfigurator.configure();
-
-		ReaderBenchView.initializeDB();
-
-		ReaderBenchView.adjustToSystemGraphics();
-		BasicConfigurator.configure();
-		processAllFolders("resources/in/blogs_Nic/1 year", "LM2", "resources/config/LSA/tasa_en",
-				"resources/config/LDA/tasa_en", Lang.eng, true, true, null, null, 0, 7);
-		// String path = "resources/in/MOOC/forum_posts&comments";
-		// SerialCorpusAssessment.processCorpus(path,
-		// "resources/config/LSA/tasa_lak_en",
-		// "resources/config/LDA/tasa_lak_en", Lang.eng, true, false, null,
-		// null, true);
-		// Long startDate = 1383235200L;
-		// Long startDate = 1383843600L; new Date(startDate * 1000)
-		// Community.processDocumentCollection(path, false, null, null, 0, 7);
 	}
 }
